Adding Testing Folder
This folder includes the initial tests used to check the end-to-end functionality of the model for the Utopia and Test System datasets.  The sqlite databases are currently included in the commit for ease of use, but they should be removed from VCS eventually as any test run modifies them with outputs and will cause repository bloat.