Emission Cost Computation into Output
Emission Costs (discounted and undiscounted) are now captured in the cost table.  First check shows a good match on simple test case with Utopia and 1 emission in 1 period:  total in table matches objective value.

Also removed rounding off in the output table and adjusted the test value in the test case for myopic to accommodate that.