Adjusted LaTeX for Reserve Margin Constraint
The Reserve Margin Constrain formula is long and had a multiline environment clause in it that was hanging up the make command.  That environment clause is removed.  The format of the equation is not that great in the pdf or html.  It could use some love from someone with experience aligning long equations or such....  But it is functional now.  make html and make latexpdf both appear to work fine.